sbllm_repo/requirement.txt
# Core dependencies
jsonlines>=3.1.0
numpy>=1.24.0
tqdm>=4.65.0
pandas>=2.0.0

# LLM API clients
openai>=0.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
nltk>=3.8.1

# Text processing and retrieval
rank_bm25>=0.2.2
editdistance>=0.6.0
fastbm25>=0.0.1

# Code analysis
tree-sitter==0.21.3
tree-sitter-cpp>=0.21.0
codebleu==0.6.0

# Vector embeddings (optional, for hybrid retrieval)
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0